billing_service/tests/functional/requirements.txt
pytest==8.3.5
pytest-asyncio==0.25.3
requests==2.32.3
aiohttp==3.11.13
redis==5.0.4
pydantic-settings==2.8.0
sqlalchemy[asyncio]==2.0.39
python-dotenv==1.1.0
asyncpg==0.30.0
Werkzeug==3.1.3
alembic==1.15.1
psycopg[binary]==3.2.9
PyJWT==2.10.1
passlib==1.7.4
backoff==2.2.1
Authlib==1.5.2
starlette==0.46.2
httpx==0.28.1
fastapi==0.115.12
orjson==3.10.18
pytest==8.3.5
pytest-asyncio==0.25.3
pydantic-settings==2.8.0
pytest-watch==4.2.0
uvicorn==0.34.3
psycopg2-binary==2.9.9
yookassa==3.5.0
pydash==8.0.5